# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Bill:
    properties:
      Amount:
        description: Amount holds the value of the "Amount" field.
        type: string
      Date:
        description: Date holds the value of the "Date" field.
        type: string
      edges:
        $ref: '#/definitions/ent.BillEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BillQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.BillEdges:
    properties:
      officer:
        $ref: '#/definitions/ent.Financier'
        description: Officer holds the value of the officer edge.
        type: object
      paytype:
        $ref: '#/definitions/ent.Paytype'
        description: Paytype holds the value of the paytype edge.
        type: object
      treatment:
        $ref: '#/definitions/ent.Unpaybill'
        description: Treatment holds the value of the treatment edge.
        type: object
    type: object
  ent.Department:
    properties:
      department:
        description: Department holds the value of the "department" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DepartmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DepartmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
...
  ent.Doctorinfo:
    properties:
      doctorname:
        description: Doctorname holds the value of the "doctorname" field.
        type: string
      doctorsurname:
        description: Doctorsurname holds the value of the "doctorsurname" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DoctorinfoEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DoctorinfoQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      licensenumber:
        description: Licensenumber holds the value of the "licensenumber" field.
        type: string
      telephonenumber:
        description: Telephonenumber holds the value of the "telephonenumber" field.
        type: string
    type: object
  ent.DoctorinfoEdges:
    properties:
      department:
        $ref: '#/definitions/ent.Department'
        description: Department holds the value of the department edge.
        type: object
      educationlevel:
        $ref: '#/definitions/ent.Educationlevel'
        description: Educationlevel holds the value of the educationlevel edge.
        type: object
      officeroom:
        $ref: '#/definitions/ent.Officeroom'
        description: Officeroom holds the value of the officeroom edge.
        type: object
      prename:
        $ref: '#/definitions/ent.Prename'
        description: Prename holds the value of the prename edge.
        type: object
      treatment:
        description: Treatment holds the value of the treatment edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
...
  ent.Financier:
    properties:
      edges:
        $ref: '#/definitions/ent.FinancierEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FinancierQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.FinancierEdges:
    properties:
      bills:
        description: Bills holds the value of the bills edge.
        items:
          $ref: '#/definitions/ent.Bill'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
...
  ent.Medicalrecordstaff:
    properties:
      Name:
        description: Name holds the value of the "Name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.MedicalrecordstaffEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MedicalrecordstaffQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.MedicalrecordstaffEdges:
    properties:
      medicalrecordstaffPatientrights:
        description: MedicalrecordstaffPatientrights holds the value of the MedicalrecordstaffPatientrights
          edge.
        items:
          $ref: '#/definitions/ent.Patientrights'
        type: array
      patientrecord:
        description: Patientrecord holds the value of the patientrecord edge.
        items:
          $ref: '#/definitions/ent.Patientrecord'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
  ent.Nurse:
    properties:
      edges:
        $ref: '#/definitions/ent.NurseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the NurseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      nursinglicense:
        description: Nursinglicense holds the value of the "nursinglicense" field.
        type: string
      position:
        description: Position holds the value of the "position" field.
        type: string
    type: object
  ent.NurseEdges:
    properties:
      historytaking:
        description: Historytaking holds the value of the historytaking edge.
        items:
          $ref: '#/definitions/ent.Historytaking'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
...
  ent.Patientrights:
    properties:
      PermissionDate:
        description: PermissionDate holds the value of the "PermissionDate" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PatientrightsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PatientrightsQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      insurance_id:
        type: integer
      patientrightstype_id:
        type: integer
    type: object
  ent.PatientrightsEdges:
    properties:
      patientrightsInsurance:
        $ref: '#/definitions/ent.Insurance'
        description: PatientrightsInsurance holds the value of the PatientrightsInsurance
          edge.
        type: object
      patientrightsMedicalrecordstaff:
        $ref: '#/definitions/ent.Medicalrecordstaff'
        description: PatientrightsMedicalrecordstaff holds the value of the PatientrightsMedicalrecordstaff
          edge.
        type: object
      patientrightsPatientrecord:
        $ref: '#/definitions/ent.Patientrecord'
        description: PatientrightsPatientrecord holds the value of the PatientrightsPatientrecord
          edge.
        type: object
      patientrightsPatientrightstype:
        $ref: '#/definitions/ent.Patientrightstype'
        description: PatientrightsPatientrightstype holds the value of the PatientrightsPatientrightstype
          edge.
        type: object
    type: object
...
  ent.Paytype:
    properties:
      edges:
        $ref: '#/definitions/ent.PaytypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PaytypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      paytype:
        description: Paytype holds the value of the "paytype" field.
        type: string
    type: object
  ent.PaytypeEdges:
    properties:
      bills:
        description: Bills holds the value of the bills edge.
        items:
          $ref: '#/definitions/ent.Bill'
        type: array
    type: object
...
  ent.Treatment:
    properties:
      datetime:
        description: Datetime holds the value of the "datetime" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TreatmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TreatmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      treatment:
        description: Treatment holds the value of the "treatment" field.
        type: string
    type: object
  ent.TreatmentEdges:
    properties:
      doctorinfo:
        $ref: '#/definitions/ent.Doctorinfo'
        description: Doctorinfo holds the value of the doctorinfo edge.
        type: object
      patientrecord:
        $ref: '#/definitions/ent.Patientrecord'
        description: Patientrecord holds the value of the patientrecord edge.
        type: object
      typetreatment:
        $ref: '#/definitions/ent.Typetreatment'
        description: Typetreatment holds the value of the typetreatment edge.
        type: object
      unpaybills:
        $ref: '#/definitions/ent.Unpaybill'
        description: Unpaybills holds the value of the unpaybills edge.
        type: object
    type: object
  ent.Typetreatment:
    properties:
      edges:
        $ref: '#/definitions/ent.TypetreatmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TypetreatmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      type:
        description: Type holds the value of the "type" field.
        type: string
    type: object
...
  ent.Unpaybill:
    properties:
      Status:
        description: Status holds the value of the "Status" field.
        type: string
      edges:
        $ref: '#/definitions/ent.UnpaybillEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UnpaybillQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.UnpaybillEdges:
    properties:
      bills:
        $ref: '#/definitions/ent.Bill'
        description: Bills holds the value of the bills edge.
        type: object
      treatment:
        $ref: '#/definitions/ent.Treatment'
        description: Treatment holds the value of the treatment edge.
        type: object
    type: object
  ent.User:
    properties:
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      financier:
        $ref: '#/definitions/ent.Financier'
        description: Financier holds the value of the financier edge.
        type: object
      historytaking:
        $ref: '#/definitions/ent.Nurse'
        description: Historytaking holds the value of the historytaking edge.
        type: object
      medicalrecordstaff:
        $ref: '#/definitions/ent.Medicalrecordstaff'
        description: Medicalrecordstaff holds the value of the medicalrecordstaff
          edge.
        type: object
      user2doctorinfo:
        $ref: '#/definitions/ent.Doctorinfo'
        description: User2doctorinfo holds the value of the user2doctorinfo edge.
        type: object
      userPatientrights:
        $ref: '#/definitions/ent.Patientrights'
        description: UserPatientrights holds the value of the UserPatientrights edge.
        type: object
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API
  version: "1.0"
paths:
...
  /bills:
    get:
      description: list bill entities
      operationId: list-bill
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Bill'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List bill entities
    post:
      consumes:
      - application/json
      description: Create bill
      operationId: create-bill
      parameters:
      - description: Bill entity
        in: body
        name: bill
        required: true
        schema:
          $ref: '#/definitions/ent.Bill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Bill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create bill
  /bills/{id}:
    delete:
      description: get bill by ID
      operationId: delete-bill
      parameters:
      - description: Bill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a bill entity by ID
    get:
      description: get bill by ID
      operationId: get-bill
      parameters:
      - description: Bill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Bill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a bill entity by ID
...
  /unpaybills:
    get:
      description: list unpaybill entities
      operationId: list-unpaybill
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Unpaybill'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List unpaybill entities
    post:
      consumes:
      - application/json
      description: Create unpaybill
      operationId: create-unpaybill
      parameters:
      - description: Unpaybill entity
        in: body
        name: unpaybill
        required: true
        schema:
          $ref: '#/definitions/ent.Unpaybill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Unpaybill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create unpaybill
  /unpaybills/{id}:
    delete:
      description: get unpaybill by ID
      operationId: delete-unpaybill
      parameters:
      - description: Unpaybill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a unpaybill entity by ID
    get:
      description: get unpaybill by ID
      operationId: get-unpaybill
      parameters:
      - description: Unpaybill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Unpaybill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a unpaybill entity by ID
    put:
      consumes:
      - application/json
      description: update unpaybill by ID
      operationId: update-unpaybill
      parameters:
      - description: Unpaybill ID
        in: path
        name: id
        required: true
        type: integer
      - description: Unpaybill entity
        in: body
        name: unpaybill
        required: true
        schema:
          $ref: '#/definitions/ent.Unpaybill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Unpaybill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a unpaybill entity by ID
  /users:
    get:
      description: list user entities
      operationId: list-user
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List user entities
    post:
      consumes:
      - application/json
      description: Create user
      operationId: create-user
      parameters:
      - description: User entity
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create user
  /users/{id}:
    delete:
      description: get user by ID
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a user entity by ID
    get:
      description: get user by ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user entity by ID
    put:
      consumes:
      - application/json
      description: update user by ID
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User entity
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a user entity by ID
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"